Add support for synchronization tokens
The synchronization across user devices is not started automatically
anymore. Each client will send a synchronization request at the end of
the process. Each request has a random token, which is used for
identifing user devices.

The client will not start a synchronization is the received token
is the same that was sent.